# pkg/api/v1/transactional/transactional.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/v1/transactional/transactional.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "TransactionalService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/add": {
      "post": {
        "summary": "Добавить пользователя",
        "operationId": "TransactionalService_AddUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/transactionalAddUserResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/transactionalAddUserRequest"
            }
          }
        ],
        "tags": [
          "TransactionalService"
        ]
      }
    },
    "/v1/balance/{clientId}": {
      "get": {
        "summary": "Получить баланс",
        "operationId": "TransactionalService_Balance",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/transactionalBalanceResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uint64"
          }
        ],
        "tags": [
          "TransactionalService"
        ]
      }
    },
    "/v1/deposit": {
      "put": {
        "summary": "Пополнить",
        "operationId": "TransactionalService_Deposit",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/transactionalDepositRequest"
            }
          }
        ],
        "tags": [
          "TransactionalService"
        ]
      }
    },
    "/v1/withdraw": {
      "put": {
        "summary": "Снять",
        "operationId": "TransactionalService_Withdraw",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/transactionalWithdrawRequest"
            }
          }
        ],
        "tags": [
          "TransactionalService"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "typeUrl": {
          "type": "string"
        },
        "value": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "transactionalAddUserRequest": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "balance": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "transactionalAddUserResponse": {
      "type": "object",
      "properties": {
        "clientId": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "transactionalBalanceResponse": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "balance": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "transactionalDepositRequest": {
      "type": "object",
      "properties": {
        "clientId": {
          "type": "string",
          "format": "uint64"
        },
        "amount": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "transactionalWithdrawRequest": {
      "type": "object",
      "properties": {
        "clientId": {
          "type": "string",
          "format": "uint64"
        },
        "amount": {
          "type": "string",
          "format": "uint64"
        }
      }
    }
  }
}